Add files via upload
Face recognition is a step further than face detection. In face detection, we only detect the location of the human face in an image but in face recognition, we make a system that can identify humans

It can also be stated as "Face recognition is a broad challenge of verifying or identifying people in pictures or videos. Big tech giants are still working to make a faster and more accurate face recognition model."

Project stages

Face Detection: Locate faces and draw bounding boxes around faces and keep the coordinates of bounding boxes.

Face Alignments: Normalize the faces to be consistent with the training database.

Feature Extraction: Extract features of faces that will be used for training and recognition tasks.